fix: use full-length commit SHA1
Prior to this change, the tools I use, namely [Nix] and [naersk],
imports and parses the Cargo.lock file in order to reproducibly
determine the set of crates that need to be downloaded.  During this
process it appears that naersk does additional validation of crates' rev
and results in the following error:

    error: hash '7010efd' has wrong length for hash type 'sha1'

This change simply specifies the full hash for the crate that causes the
error, which is sufficient to resolve the error and also has no impact
on other build tools.

[Nix]: https://nixos.org/
[naersk]: https://github.com/nix-community/naersk